Final commit, program doesn't work like expected

+++ Vehicles/Vehicle.m
@@ -35,7 +35,7 @@
         degrees = degrees % degreesInACircle;
     }
     
-    return [NSString stringWithFormat:@"Turn %d degrees.", degrees];
+    return [NSString stringWithFormat:@"Turn %ld degrees.", (long)degrees];
 }
 
 -(NSString *)changeGears:(NSString *)newGearName
@@ -51,7 +51,7 @@
 #pragma mark - Convenience Methods
 -(NSString *)vehicleTitleString
 {
-    return [NSString stringWithFormat:@"%d %@ %@", self.modelYear, self.brandName, self.modelName];
+    return [NSString stringWithFormat:@"%ld %@ %@", (long)self.modelYear, self.brandName, self.modelName];
 }
 
 -(NSString *)vehicleDetailsString
@@ -61,9 +61,9 @@
     [basicDetailsString appendString:@"Basic vehicle details:\n\n"];
     [basicDetailsString appendFormat:@"Brand name: %@\n", self.brandName];
     [basicDetailsString appendFormat:@"Model name: %@\n", self.modelName];
-    [basicDetailsString appendFormat:@"Model year: %d\n", self.modelYear];
+    [basicDetailsString appendFormat:@"Model year: %ld\n", (long)self.modelYear];
     [basicDetailsString appendFormat:@"Power source: %@\n", self.powerSource];
-    [basicDetailsString appendFormat:@"# of wheels: %d", self.numberOfWheels];
+    [basicDetailsString appendFormat:@"# of wheels: %ld", (long)self.numberOfWheels];
     
     return basicDetailsString;
 }
